requirements.txt
# Core RAG & Embeddings
chromadb
sentence-transformers
langchain==1.2.13
langchain-community
langchain-text-splitters

# Document Loaders
pypdf
python-docx
unstructured

# Hybrid Search (BM25)
rank_bm25

# Gemini API (Google Generative AI)
google-generativeai

# OpenAI (for GPT-4o-mini planning agents)
openai

# CrewAI Multi-Agent Framework
crewai==1.12.2
crewai-tools==1.12.2

# MongoDB Async Driver
motor

# CAS PDF Parser (CAMS/KFintech)
casparser[fast]

# Demo Web Server / Frontend
fastapi
uvicorn[standard]
python-multipart

# Utilities
python-dotenv
rich
pydantic>=2.8.2
pydantic-settings
